Allow custom repository to edit where clause
